Renommage du component de statistiques.
